Conflicts:
	c# folder/C# swingame/obj/x86/Debug-Mac/MyGame.csprojResolveAssemblyReference.cache
	c# folder/C# swingame/src/GameLogic.cs
	c# folder/C# swingame/src/Model/ShipName.cs
	c# folder/C# swingame/src/Model/Tile.cs

+++ c# folder/C# swingame/src/Model/Ship.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
+
 /// <summary>
 /// A Ship has all the details about itself. For example the shipname,
 /// size, number of hits taken and the location. Its able to add tiles,
@@ -78,7 +79,7 @@ public class Ship
 		_tiles = new List<Tile>();
 
 		//gets the ship size from the enumarator
-		_sizeOfShip = _shipName;
+		_sizeOfShip = (int)_shipName;
 	}
 
 	/// <summary>
@@ -131,3 +132,4 @@ public class Ship
 		_direction = direction;
 	}
 }
+
